+++ extensions/bike/src/append-from-clipboard.ts
@@ -13,33 +13,26 @@ export default function main() {
     setRanScript(true);
 
     // Get lines of text from the clipboard
-    Promise.resolve(Clipboard.readText()).then((text) => {
+    Clipboard.readText().then((text) => {
       const lines = text?.split("\n");
       const clipboard_lines = lines?.map(
         (line: string) => '"' + line.replaceAll("\\", "\\\\").replaceAll('"', '\\"') + '"'
       );
 
       // Run script
-      Promise.resolve(
-        runAppleScript(`tell application "Bike"
+      runAppleScript(`tell application "Bike"
         activate
         -- Get the most recent document
-        set theDoc to document 0
+        set theDoc to document 1
 
         -- Get lines of clipboard content
         set docData to {${clipboard_lines}}
 
-        -- Add the clipboard content to the beginning of the document
+        -- Add the clipboard content to the end of the document
         repeat with lineItem in docData
           tell theDoc to make new row with properties {name: lineItem}
         end repeat
-      end tell`)
-      );
+      end tell`).then(() => showHUD("Appended To Current Document").then(() => popToRoot()));
     });
-
-    showHUD("Appended to Bike");
-
-    // Close the Raycast window
-    Promise.resolve(popToRoot());
   }
 }