+++ extensions/bike/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Bike Changelog
 
+## [Search Rows & Copy Row Link] - 2022-01-25
+- Add "Copy Row URL" command
+- Add "Search Rows" command
+- Rename "Copy URL" to "Copy Document URL"
+- Rename "Close Other Bikes" to "Close Other Documents"
+- Rename "Force Close Other Bikes" to "Force Close Other Documents"
+
 ## [Initial Version] - 2022-09-05